requirements.txt
pymongo==3.11.4
dnspython==2.1.0
fastapi==0.66.0
